+++ .devcontainer/dev/setup-remote-vm.sh
@@ -9,7 +9,7 @@ set -eo pipefail
 BUILDARCH=$([ "$(uname -i)" = "aarch64" ] && echo "arm64" || echo "amd64")
 HOSTNAME="knfsd-dev-ec2"
 USERNAME="ubuntu"
-VERSION="1.1.0-alpha.8"
+VERSION="1.1.0-alpha.9"
 
 ## set env vars for build env only
 export HOME=/home/${USERNAME}
@@ -128,9 +128,9 @@ KNFSD_BATS_CORE_VERSION=1.12.0
 # https://github.com/psf/black/releases
 KNFSD_BLACK_VERSION=25.1.0
 # https://github.com/boto/boto3/tags
-KNFSD_BOTO3_VERSION=1.40.21
+KNFSD_BOTO3_VERSION=1.40.30
 # https://hub.docker.com/r/bridgecrew/checkov/tags
-KNFSD_CHECKOV_VERSION=3.2.469
+KNFSD_CHECKOV_VERSION=3.2.471
 # https://github.com/codespell-project/codespell/releases
 KNFSD_CODESPELL_VERSION=2.4.1
 # https://github.com/editorconfig-checker/editorconfig-checker/releases
@@ -138,19 +138,19 @@ KNFSD_EDITORCONFIG_VERSION=3.4.0
 # https://github.com/golangci/golangci-lint/releases
 KNFSD_GOLANGCI_LINT_VERSION=2.4.0
 # https://go.dev/dl/
-KNFSD_GOLANG_VERSION=1.25.0
+KNFSD_GOLANG_VERSION=1.25.1
 # https://github.com/python/mypy/tags
-KNFSD_MYPY_VERSION=1.17.1
+KNFSD_MYPY_VERSION=1.18.1
 # https://github.com/hashicorp/packer/releases
-KNFSD_PACKER_VERSION=1.14.1
+KNFSD_PACKER_VERSION=1.14.2
 # https://github.com/pre-commit/pre-commit/releases
 KNFSD_PRECOMMIT_VERSION=4.3.0
 # https://pypi.org/project/psycopg/
-KNFSD_PSYCOPG_VERSION=3.2.9
+KNFSD_PSYCOPG_VERSION=3.2.10
 # https://github.com/pylint-dev/pylint/tags
 KNFSD_PYLINT_VERSION=3.3.8
 # https://github.com/semgrep/semgrep/releases
-KNFSD_SEMGREP_VERSION=1.134.0
+KNFSD_SEMGREP_VERSION=1.136.0
 # https://pypi.org/project/shellcheck-py/
 KNFSD_SHELLCHECK_PY_VERSION=0.11.0.1
 # https://github.com/mvdan/sh/releases
@@ -158,13 +158,13 @@ KNFSD_SHFMT_VERSION=3.12.0
 # https://github.com/hashicorp/terraform/releases
 KNFSD_TERRAFORM_VERSION=1.2.9
 # https://github.com/gruntwork-io/terragrunt/releases
-KNFSD_TERRAGRUNT_VERSION=0.86.2
+KNFSD_TERRAGRUNT_VERSION=0.87.2
 # https://github.com/terraform-linters/tflint/releases
 KNFSD_TFLINT_VERSION=0.59.1
 # https://github.com/aquasecurity/tfsec/releases
 KNFSD_TFSEC_VERSION=1.28.14
 # https://github.com/aquasecurity/trivy/releases
-KNFSD_TRIVY_VERSION=0.65.0
+KNFSD_TRIVY_VERSION=0.66.0
 
 ## install golang, delete empty lines and lines containing PATH= in /etc/environment
 curl -fsSL "https://dl.google.com/go/go${KNFSD_GOLANG_VERSION}.linux-${BUILDARCH}.tar.gz" -o "/tmp/go${KNFSD_GOLANG_VERSION}.linux-${BUILDARCH}.tar.gz" \