+++ examples/fsx-zfs/main.tf
@@ -8,7 +8,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 6.11.0"
+      version = "~> 6.13.0"
     }
   }
 }
@@ -155,12 +155,13 @@ resource "aws_security_group" "fsx_sg" {
 module "proxy" {
   source                  = "../../deployment/terraform-module-knfsd"
   SUBNET                  = var.SUBNET
-  KNFSD_NODES             = 1
+  KNFSD_NODES             = var.KNFSD_NODES
   PROXY_AMI               = var.PROXY_AMI
   INSTANCE_TAGS           = { "knfsd-file-cache:examples" = "fsx-zfs" }
   PROXY_BASENAME          = var.PROXY_BASENAME
   TRAFFIC_MODE            = "dns_round_robin"
   KEY_NAME                = var.KEY_NAME
+  INSTANCE_TYPE           = var.INSTANCE_TYPE
   FSID_MODE               = var.FSID_MODE
   EXPORT_HOST_AUTO_DETECT = aws_fsx_openzfs_file_system.zfs.dns_name # Detect exports from the source filer via "showmount -e <SOURCE_FILER_DNS_NAME>"
   EXPORT_OPTIONS          = "insecure"                               # Override the default "secure" option with "insecure" (required for "showmount" auto-discovery by clients)